fixing some bugs on Tournament system

+++ domain/src/main/java/com/bueno/domain/usecases/tournament/dtos/MatchDTO.java
@@ -1,8 +1,5 @@
 package com.bueno.domain.usecases.tournament.dtos;
 
-import com.bueno.domain.entities.tournament.Match;
-
-import java.util.Comparator;
 import java.util.UUID;
 
 public record MatchDTO(UUID id,
@@ -11,15 +8,15 @@ public record MatchDTO(UUID id,
                        String p2Name,
                        boolean available,
                        String winnerName,
-                       int p1Score,
-                       int p2Score,
+                       long p1Score,
+                       long p2Score,
                        UUID next) {
 
     @Override
     public boolean equals(Object o) {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
-
+        
         MatchDTO matchDTO = (MatchDTO) o;
         return id.equals(matchDTO.id);
     }
